docs(grpc): deprecate DeployProcess rpc
The DeployProcess RPC has been deprecated since 8. Users should switch
to the more flexible DeployResource RPC. Of course the documentation
should describe this.

To fit it in this page, a new Deprecated RPCs heading is added. The
deprecated RPC has been moved to it and a deprecation note has been
added explaining since when it was deprecated and that is has been
replaced by the DeployResource RPC.